// SGUEES-SPA/src/app/pages/Shop/com-cuadro-comparativo/models/com-cuadro-comparativo.ts
export interface ComCuadroComparativo {
	CORR_EMPRESA: number;
	ANIO_PERIODO: number;
	CORR_CUADRO_COMPARATIVO: number;
  NUMERO_CUADRO_COMPARA: string;
	FECHA_CUADRO_COMPARATIVO: Date;
	ESTADO_CUADRO_COMPARATIVO: string;
  NOMBRE_ESTADO_CUADRO_COMPARATIVO: string;
	USUARIO_CUADRO_COMPARATIVO: string;
	OBSERVACIONES: string;
	ES_MEJOR_PRECIO: boolean;
	TIENE_CREDITO_30_DIAS: boolean;
	TIENE_BUEN_SOPORTE_TECNICO: boolean;
	TIENE_BUENA_CALIDAD_PRODUCTO: boolean;
	TIENE_MEJOR_TIEMPO_ENTREGA: boolean;
	BRINDA_BUENA_EXPERIENCIA_PROVEEDOR: boolean;
	ES_PROVEEDOR_UNICO: boolean;
	EXISTE_OTRA_RAZON: boolean;
	NOMBRE_OTRA_RAZON: string;
	ANIO_PERIODO_SOLI_COTI: number;
	CORR_SOLI_COTIZACION: number;
	USUARIO_CREA: string;
	FECHA_CREA: Date;
	ESTACION_CREA: string;
	USUARIO_ACTU: string;
	FECHA_ACTU: Date;
	ESTACION_ACTU: string;
  NOMBRE_PROVEEDOR: string;
  NUMERO_SOLI_COTIZACION: string;
  NUMERO_SOLI_COMPRA: string;
  NUMERO_COTIZACION: string;
}